+++ packages/constructor/src/constructors/TableRecord/TableRecord.ts
@@ -5,8 +5,10 @@ import {
   isString
 } from '@dxtmisha/functional'
 
+import { ClientOnlyInclude } from '../../classes/ClientOnlyInclude'
 import { TableHeaderItemInclude } from '../TableHeaderItem'
 import { TableItemInclude } from '../TableItem'
+import { TableRecordLazy } from './TableRecordLazy'
 
 import type { TableRecordComponents, TableRecordEmits, TableRecordSlots } from './types'
 import type { TableRecordProps } from './props'
@@ -24,6 +26,11 @@ export class TableRecord {
   /** Table items include instance / Экземпляр включения ячеек таблицы */
   readonly tableItem: TableItemInclude
 
+  /** Client-only rendering include / Подключение рендеринга только на клиенте */
+  readonly clientOnly: ClientOnlyInclude
+  /** Table record lazy loading handler instance / Экземпляр обработчика ленивой загрузки записи таблицы */
+  readonly lazy: TableRecordLazy
+
   /**
    * Constructor
    * @param props input properties / входные свойства
@@ -35,8 +42,10 @@ export class TableRecord {
    * @param slots object for working with slots / объект для работы со слотами
    * @param emits callback function triggered on events / функция обратного вызова, запускаемая при событиях
    * @param constructors optional class constructor overrides / необязательные переопределения конструкторов классов
+   * @param constructors.ClientOnlyIncludeConstructor class for creating a client only include / класс для создания включения рендеринга только на клиенте
    * @param constructors.TableHeaderItemIncludeConstructor class for creating a table header item include / класс для создания включения ячейки шапки таблицы
    * @param constructors.TableItemIncludeConstructor class for creating a table item include / класс для создания включения ячейки таблицы
+   * @param constructors.TableRecordLazyConstructor class for creating a table record lazy loading handler / класс для создания обработчика ленивой загрузки записи таблицы
    */
   constructor(
     protected readonly props: TableRecordProps,
@@ -48,13 +57,17 @@ export class TableRecord {
     protected readonly slots?: TableRecordSlots,
     protected readonly emits?: ConstrEmit<TableRecordEmits>,
     constructors: {
+      ClientOnlyIncludeConstructor?: typeof ClientOnlyInclude
       TableHeaderItemIncludeConstructor?: typeof TableHeaderItemInclude
       TableItemIncludeConstructor?: typeof TableItemInclude
+      TableRecordLazyConstructor?: typeof TableRecordLazy
     } = {}
   ) {
     const {
+      ClientOnlyIncludeConstructor = ClientOnlyInclude,
       TableHeaderItemIncludeConstructor = TableHeaderItemInclude,
-      TableItemIncludeConstructor = TableItemInclude
+      TableItemIncludeConstructor = TableItemInclude,
+      TableRecordLazyConstructor = TableRecordLazy
     } = constructors
 
     this.tableHeaderItem = new TableHeaderItemIncludeConstructor(
@@ -73,6 +86,14 @@ export class TableRecord {
       components,
       undefined,
       slots
+    )
+
+    this.clientOnly = new ClientOnlyIncludeConstructor()
+    this.lazy = new TableRecordLazyConstructor(
+      props,
+      element,
+      className,
+      this.clientOnly
     )
   }